+++ lib/src/cache/service/platform_cache_helper.dart
@@ -1,0 +1,3 @@
+export 'platform_cache_helper_stub.dart'
+    if (dart.library.io) 'platform_cache_helper_io.dart'
+    if (dart.library.html) 'platform_cache_helper_web.dart';